+++ tests/Test.hs
@@ -30,28 +30,27 @@ main = hspec $ do
     propJSONEq (Proxy :: Proxy Suggested)
     propJSONApproxEq (Proxy :: Proxy PostDates)
 
-  describe "decodeJSONResponse: handle parse failures" $ do
-    it "object parse failure" $
-        let noteJson = "FAIL"
-        in case (decodeJSONResponse noteJson) of
-          Left PinboardError{..} -> errorType == ParseFailure
-          Right Note{..} -> False
-    it "field parse failure" $
-        let noteJson = "{\"length\":0,\"hash\":\"\",\"text_FAIL\":\"\",\"updated_at\":\"1864-05-09 13:50:53\",\"created_at\":\"1864-05-09 18:21:35\",\"id\":\"\",\"title\":\"\"}"
-        in case (decodeJSONResponse noteJson) of
-          Left PinboardError{..} -> errorType == ParseFailure
-          Right Note{..} -> False
-    it "value parse failure" $
-        let noteJson = "{\"length\":FAIL,\"hash\":\"\",\"text_FAIL\":\"\",\"updated_at\":\"1864-05-09 13:50:53\",\"created_at\":\"1864-05-09 18:21:35\",\"id\":\"\",\"title\":\"\"}"
-        in case (decodeJSONResponse noteJson) of
-          Left PinboardError{..} -> errorType == ParseFailure
-          Right Note{..} -> False
-    it "time parse failure" $
-        let noteJson = "{\"length\":0,\"hash\":\"\",\"text\":\"\",\"updated_at\":\"FAIL-05-09 13:50:53\",\"created_at\":\"1864-05-09 18:21:35\",\"id\":\"\",\"title\":\"\"}"
-        in case (decodeJSONResponse noteJson) of
-          Left PinboardError{..} -> errorType == ParseFailure
-          Right Note{..} -> False
-
+    describe "decodeJSONResponse: handle parse failures" $ do
+        it "object parse failure" $
+            let noteJson = "FAIL"
+            in case decodeJSONResponse noteJson of
+            Left PinboardError{..} -> errorType == ParseFailure
+            Right Note{..} -> False
+        it "field parse failure" $
+            let noteJson = "{\"length\":0,\"hash\":\"\",\"text_FAIL\":\"\",\"updated_at\":\"1864-05-09 13:50:53\",\"created_at\":\"1864-05-09 18:21:35\",\"id\":\"\",\"title\":\"\"}"
+            in case decodeJSONResponse noteJson of
+            Left PinboardError{..} -> errorType == ParseFailure
+            Right Note{..} -> False
+        it "value parse failure" $
+            let noteJson = "{\"length\":FAIL,\"hash\":\"\",\"text_FAIL\":\"\",\"updated_at\":\"1864-05-09 13:50:53\",\"created_at\":\"1864-05-09 18:21:35\",\"id\":\"\",\"title\":\"\"}"
+            in case decodeJSONResponse noteJson of
+            Left PinboardError{..} -> errorType == ParseFailure
+            Right Note{..} -> False
+        it "time parse failure" $
+            let noteJson = "{\"length\":0,\"hash\":\"\",\"text\":\"\",\"updated_at\":\"FAIL-05-09 13:50:53\",\"created_at\":\"1864-05-09 18:21:35\",\"id\":\"\",\"title\":\"\"}"
+            in case decodeJSONResponse noteJson of
+            Left PinboardError{..} -> errorType == ParseFailure
+            Right Note{..} -> False
 
 pinboardParseFailure :: Selector PinboardError
 pinboardParseFailure e = errorType e == ParseFailure