+++ changelog.md
@@ -1,7 +1,10 @@
 __v0.9.8__
 
 use safe-exception
+
 avoid use of 'error' to communicate failure when parsing json
+
+includes a few changes to the types/api in Pinboard.Client
 
 __v0.9.7__